# swagger.yaml
openapi: 3.0.0
info:
  title: Go Fit API
  version: 1.0.0
  description: Comprehensive API documentation for the Go Fit App, providing access to features such as user management, class bookings, orders, and more. This API is designed to support both trainers and clients, enabling seamless interactions with the platform's services.
servers:
  - url: http://localhost:9876
    description: Server
paths:
...
  /user/all:
    get:
      tags:
        - User
      summary: Get all users with role "client"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        name:
                          type: string
                        email:
                          type: string
                        password:
                          type: string
                        phone:
                          type: number
                        country:
                          type: string
                        sex:
                          type: string
                        role:
                          type: string
                          enum: [trainer, client]
                        age:
                          type: number
                        height:
                          type: number
                        weight:
                          type: number
                        healthProblem:
                          type: array
                          items:
                            type: string
                        classes:
                          type: array
                          items:
                            type: string
                        createdDate:
                          type: string
                        createdTime:
                          type: string
        "400":
          description: Error occurred
...
  /user/singletrainer/{id}:
    get:
      tags:
        - User
      summary: Get a single trainer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trainer data fetched
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  trainer:
                    type: object
                    properties:
                      _id:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
                      password:
                        type: string
                      phone:
                        type: number
                      country:
                        type: string
                      sex:
                        type: string
                      role:
                        type: string
                        enum: [trainer, client]
                      age:
                        type: number
                      height:
                        type: number
                      weight:
                        type: number
                      healthProblem:
                        type: array
                        items:
                          type: string
                      classes:
                        type: array
                        items:
                          type: string
                      createdDate:
                        type: string
                      createdTime:
                        type: string
        "400":
          description: Error occurred
  /user/{id}:
    get:
      tags:
        - User
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User data fetched
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    type: object
                    properties:
                      _id:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
                      password:
                        type: string
                      phone:
                        type: number
                      country:
                        type: string
                      sex:
                        type: string
                      role:
                        type: string
                        enum: [trainer, client]
                      age:
                        type: number
                      height:
                        type: number
                      weight:
                        type: number
                      healthProblem:
                        type: array
                        items:
                          type: string
                      classes:
                        type: array
                        items:
                          type: string
                      createdDate:
                        type: string
                      createdTime:
                        type: string
        "400":
          description: Error occurred
  /user/register:
    post:
      tags:
        - User
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                phone:
                  type: number
                country:
                  type: string
                sex:
                  type: string
                role:
                  type: string
                  enum: [trainer, client]
                age:
                  type: number
                height:
                  type: number
                weight:
                  type: number
                healthProblem:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: User registered
        "400":
          description: User already exists or other error
        "401":
          description: Server error
  /user/login:
    post:
      tags:
        - User
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
        "400":
          description: User not found
        "401":
          description: Incorrect password
  /user/update/{id}:
    patch:
      tags:
        - User
      summary: Update user data
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                phone:
                  type: number
                country:
                  type: string
                sex:
                  type: string
                role:
                  type: string
                  enum: [trainer, client]
                age:
                  type: number
                height:
                  type: number
                weight:
                  type: number
                healthProblem:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: User updated
        "400":
          description: Error occurred
  /user/delete/{id}:
    delete:
      tags:
        - User
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
        "400":
          description: Error occurred
...
  /order/user/{id}:
    get:
      tags:
        - Booking
      summary: Get all orders of a user by user ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID
      responses:
        200:
          description: Orders fetched
        400:
          description: Error occurred
...